requirements.txt
colorama
paramiko
polars
pyserial
pyyaml
requests
schedule
sockslib
paho-mqtt
numpy
